+++ packages/katana_listenables_builder/lib/value/class_value.dart
@@ -25,7 +25,8 @@ class ClassValue {
     });
     existChangeNotifierMixin = element.mixins.any(
       (element) =>
-          element.getDisplayString().trimString("?") == "ChangeNotifier",
+          element.getDisplayString(withNullability: true).trimString("?") ==
+          "ChangeNotifier",
     );
     parameters = contstuctor.parameters.where((e) => e.name != "key").map((e) {
       return ParamaterValue(e);