+++ packages/masamune_annotation/lib/masamune_annotation.dart
@@ -21,3 +21,4 @@ part 'src/search_param.dart';
 part 'src/json_param.dart';
 part 'src/google_spread_sheet_data_source.dart';
 part 'src/model_permission_query.dart';
+part 'src/param_comment.dart';