+++ packages/masamune_annotation/lib/src/collection_model_path.dart
@@ -138,6 +138,8 @@ class CollectionModelPath {
     this.endpoint,
     this.permission,
     this.mirrorPermission,
+    this.docsPath = "documents/docs",
+    this.comment,
   });
 
   /// Path for collection.
@@ -159,6 +161,16 @@ class CollectionModelPath {
   ///
   /// APIにアクセスするためのエンドポイント。
   final String? endpoint;
+
+  /// The path (relative path) to which the document will be generated.
+  ///
+  /// ドキュメントを生成する際の生成先のパス（相対パス）。
+  final String docsPath;
+
+  /// Comments to be assigned to the collection.
+  ///
+  /// コレクションに付与するコメント。
+  final String? comment;
 
   /// List to define permissions.
   ///