+++ packages/masamune_builder/lib/src/config.dart
@@ -4,3 +4,4 @@ const _documentModelPathChecker = TypeChecker.fromRuntime(DocumentModelPath);
 const _collectionModelPathChecker =
     TypeChecker.fromRuntime(CollectionModelPath);
 const _freezedChecker = TypeChecker.fromRuntime(Freezed);
+const _deprecatedChecker = TypeChecker.fromRuntime(Deprecated);