+++ packages/masamune_builder/lib/value/parameter_value.dart
@@ -4,6 +4,7 @@ const _searchParamChecker = TypeChecker.fromRuntime(SearchParam);
 const _refParamChecker = TypeChecker.fromRuntime(RefParam);
 const _jsonParamChecker = TypeChecker.fromRuntime(JsonParam);
 const _jsonKeyChecker = TypeChecker.fromRuntime(JsonKey);
+const _commentChecker = TypeChecker.fromRuntime(ParamComment);
 
 /// Parameter Value.
 ///
@@ -82,9 +83,24 @@ class ParamaterValue {
               ?.toStringValue() ??
           name;
     } else {
-      jsonKey = name;
+      jsonKey = _jsonKeyChecker
+              .firstAnnotationOfExact(element)
+              ?.getField("name")
+              ?.toStringValue() ??
+          name;
     }
+
+    comment = _commentChecker
+        .firstAnnotationOfExact(element)
+        ?.getField("comment")
+        ?.toStringValue();
+
+    deprecated = _deprecatedChecker
+        .firstAnnotationOfExact(element)
+        ?.getField("message")
+        ?.toStringValue();
   }
+
   static final _refParamRegExp = RegExp(r"^@RefParam\((.+)\)$");
   static final _adapterRegExp = RegExp(r"adapter\s*:\s*([^,\)]+),?");
 
@@ -127,6 +143,16 @@ class ParamaterValue {
   ///
   /// [isJsonSerializable]がtrueな場合のキーの名前。
   late final String? jsonKey;
+
+  /// Comment assigned to the field.
+  ///
+  /// フィールドに付与されたコメント。
+  late final String? comment;
+
+  /// If it is deprecated, the reason is described.
+  ///
+  /// 非推奨になっている場合はその理由が記述されます。
+  late final String? deprecated;
 
   @override
   String toString() {