+++ packages/masamune_lints/lib/common/masamune_limit_if_nesting.dart
@@ -33,7 +33,10 @@ class _MasamuneLimitIfNesting extends DartLintRule {
       }
 
       if (nestingLevel > _kMaxNestingCount) {
-        reporter.atNode(node, _code);
+        reporter.reportErrorForNode(
+          _code,
+          node,
+        );
       }
     });
   }