+++ packages/masamune_model_docs_builder/lib/src/functions.dart
@@ -1,0 +1,1 @@
+part of '/masamune_model_docs_builder.dart';