+++ packages/masamune_builder/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 2.17.4
+
+ - **FIX**: Fixed a notation bug. ([da9700f7](https://github.com/mathrunet/flutter_masamune/commit/da9700f70d0d16758d85e2075cae2bc5ebd93f49))
+
+## 2.17.3
+
+ - **FIX**: Deprecated can be obtained. ([e672d477](https://github.com/mathrunet/flutter_masamune/commit/e672d477f875fb4d0b16f9609a9c16d66cb13711))
+
+## 2.17.2
+
+ - **FIX**: Added the ability to add comments. ([3450dbb4](https://github.com/mathrunet/flutter_masamune/commit/3450dbb407b8142741671024ebf9dbe8f7b38f9e))
+
+## 2.17.1
+
+ - Update a dependency to the latest release.
+
+## 2.17.0
+
+ - **FEAT**: Added field to document path. ([ac1418ad](https://github.com/mathrunet/flutter_masamune/commit/ac1418adf5d09261668936c5d208af17ddd59483))
+
 ## 2.16.5
 
  - Update a dependency to the latest release.