+++ packages/masamune_deeplink_firebase/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.2.98
+
+ - Update a dependency to the latest release.
+
+## 2.2.97
+
+ - Update a dependency to the latest release.
+
 ## 2.2.96
 
  - **FIX**: Update Pakcages. ([71a244da](https://github.com/mathrunet/flutter_masamune/commit/71a244da130fb76cbed6ca88cf0128f99d258574))