+++ packages/masamune_location_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.4.24
+
+ - Update a dependency to the latest release.
+
+## 2.4.23
+
+ - Update a dependency to the latest release.
+
 ## 2.4.22
 
  - **FIX**: Update Pakcages. ([71a244da](https://github.com/mathrunet/flutter_masamune/commit/71a244da130fb76cbed6ca88cf0128f99d258574))