+++ packages/masamune_model_firestore_builder/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 2.3.4
+
+ - Update a dependency to the latest release.
+
+## 2.3.3
+
+ - Update a dependency to the latest release.
+
+## 2.3.2
+
+ - Update a dependency to the latest release.
+
+## 2.3.1
+
+ - Update a dependency to the latest release.
+
+## 2.3.0
+
+ - **FEAT**: Initial commit. ([d3c2aa55](https://github.com/mathrunet/flutter_masamune/commit/d3c2aa557524858fef5cd2df996c18c50c1e707a))
+
 ## 2.2.32
 
  - Update a dependency to the latest release.