+++ packages/masamune_module_point_ecosystem/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.5.103
+
+ - Update a dependency to the latest release.
+
+## 2.5.102
+
+ - Update a dependency to the latest release.
+
 ## 2.5.101
 
  - Update a dependency to the latest release.