+++ packages/masamune_notification_local/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.1.19
+
+ - Update a dependency to the latest release.
+
+## 2.1.18
+
+ - Update a dependency to the latest release.
+
 ## 2.1.17
 
  - **FIX**: Update Pakcages. ([71a244da](https://github.com/mathrunet/flutter_masamune/commit/71a244da130fb76cbed6ca88cf0128f99d258574))