+++ packages/masamune_purchase_stripe/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.5.93
+
+ - Update a dependency to the latest release.
+
+## 3.5.92
+
+ - Update a dependency to the latest release.
+
 ## 3.5.91
 
  - **FIX**: Update Pakcages. ([71a244da](https://github.com/mathrunet/flutter_masamune/commit/71a244da130fb76cbed6ca88cf0128f99d258574))